fix: resolve missing audio, clean up temp files, add windows exe icon, and bypass youtube 403 errors
- Fix missing audio in downloads by dynamically resolving the FFmpeg installation path at task start.
- Route all yt-dlp intermediate files (.part, thumbnails, subtitles) to a task-specific temp directory and recursively clean it up on task completion, failure, or cancellation.
- Prevent absolute outtmpl path override warning in yt-dlp by keeping outtmpl relative and specifying the absolute target directory in paths['home'].
- Fix YouTube HTTP 403 Forbidden errors by updating player_client in extractor_args to use ["web", "mweb", "android", "ios"] in downloader_tab and library_tab.
- Make unit tests fully platform-independent and compatible with Linux-based GitHub Actions CI/CD environments by resolving path checks dynamically.
- Add and register logo.ico in the PyInstaller spec configuration for the packaged Windows executable icon.
- Add Pillow dependency for handling the ICO icon generation.
- Document the automated test suite and updated assets layout in README.md.

+++ src/utils/speed_test.py
@@ -1,13 +1,16 @@
 import time
+
 import requests
 from PySide6.QtCore import QThread, Signal
+
 from .logger import get_logger
 
 logger = get_logger("speed_test")
 
+
 class SpeedTestWorker(QThread):
-    progress = Signal(int)         # Progress percent (0-100)
-    finished = Signal(float, str)   # speed in bytes/sec, formatted string
+    progress = Signal(int)  # Progress percent (0-100)
+    finished = Signal(float, str)  # speed in bytes/sec, formatted string
     error = Signal(str)
 
     def run(self):
@@ -18,16 +21,16 @@ class SpeedTestWorker(QThread):
             start_time = time.time()
             response = requests.get(url, stream=True, timeout=10)
             response.raise_for_status()
-            
-            total_length = response.headers.get('content-length')
+
+            total_length = response.headers.get("content-length")
             if total_length is None:
                 total_length = 5242880
             else:
                 total_length = int(total_length)
 
             downloaded = 0
-            chunk_size = 1024 * 32 # 32 KB chunks
-            
+            chunk_size = 1024 * 32  # 32 KB chunks
+
             for chunk in response.iter_content(chunk_size=chunk_size):
                 if not chunk:
                     continue
@@ -38,9 +41,9 @@ class SpeedTestWorker(QThread):
             duration = time.time() - start_time
             if duration <= 0:
                 duration = 0.001
-                
+
             speed_bps = downloaded / duration
-            
+
             # Format speed using simple utility
             speed_str = self._format_speed(speed_bps)
             logger.info(f"Speed test completed: {speed_str}")
@@ -55,7 +58,8 @@ class SpeedTestWorker(QThread):
             return "0 B/s"
         size_name = ("B/s", "KB/s", "MB/s", "GB/s")
         import math
-        i = int(math.floor(math.log(speed_bytes_sec, 1024)))
+
+        i = math.floor(math.log(speed_bytes_sec, 1024))
         p = math.pow(1024, i)
         s = round(speed_bytes_sec / p, 2)
         return f"{s} {size_name[i]}"
